requirements.txt
# AI-Driven Compliance Checking for Safety Procedures
# Core dependencies

# LLM and AI – Google Gemini
google-genai>=1.0.0
langchain>=0.2.0
langchain-google-genai>=2.1.0
langchain-community>=0.2.0

# Vector store and embeddings
faiss-cpu==1.9.0.post1
chromadb>=0.5.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
unstructured>=0.14.0

# Data models and validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# HTTP and API
requests>=2.32.0
httpx>=0.27.0
aiohttp>=3.9.0

# Text processing
nltk>=3.8.0
regex>=2024.5.0

# Report generation
jinja2>=3.1.0
markdown>=3.6

# CLI and utilities
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
